refactor: move OrderedMap to ../lib
